[CORE] Use PriorityQueue to hold task resource (#2992)

+++ gluten-data/src/main/java/io/glutenproject/memory/nmm/NativeMemoryManager.java
@@ -73,8 +73,8 @@ public class NativeMemoryManager implements TaskResource {
   }
 
   @Override
-  public long priority() {
-    return 0L; // lowest release priority
+  public int priority() {
+    return 0; // lowest release priority
   }
 
   @Override
